backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.35
alembic>=1.13.0
psycopg2-binary>=2.9.9
aiosqlite>=0.20.0

# Auth
pyjwt>=2.9.0
bcrypt>=4.2.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0

# AI
anthropic>=0.40.0
openai>=1.54.0

# Voice
telnyx>=2.0.0
deepgram-sdk>=3.7.7

# File processing
pypdf>=4.3.0
python-docx>=1.1.0

# Payments
stripe>=10.0.0

# Utils
httpx>=0.27.0
aiofiles>=24.1.0
cryptography>=42.0.0

# Gmail OAuth2 (email sending via Gmail API)
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.154.0
